refactor: group hints by category
