+++ src/hints/dict/allocFelt252Dict.test.ts
@@ -4,7 +4,7 @@ import { Register } from 'vm/instruction';
 import { VirtualMachine } from 'vm/virtualMachine';
 import { HintName } from 'hints/hintName';
 import { allocFelt252DictParser } from './allocFelt252Dict';
-import { OpType } from './hintParamsSchema';
+import { OpType } from '../hintParamsSchema';
 import { Felt } from 'primitives/felt';
 import { segmentArenaHandler } from 'builtins/segmentArena';
 import { Relocatable } from 'primitives/relocatable';
