+++ src/hints/dict/getCurrentAccessDelta.ts
