+++ src/hints/dict/getCurrentAccessIndex.ts
