+++ src/hints/dict/getNextDictKey.ts
