+++ src/hints/dict/getSegmentArenaIndex.ts
