+++ src/hints/dict/initSquashData.ts
