+++ src/hints/dict/shouldContinueSquashLoop.ts
