+++ src/hints/dict/shouldSkipSquashLoop.ts
