+++ src/hints/hintSchema.ts
@@ -1,17 +1,17 @@
 import { z } from 'zod';
 
 import { allocSegmentParser } from './allocSegment';
-import { testLessThanParser } from './testLessThan';
-import { allocFelt252DictParser } from './allocFelt252Dict';
-import { getSegmentArenaIndexParser } from './getSegmentArenaIndex';
-import { felt252DictEntryInitParser } from './felt252DictEntryInit';
-import { felt252DictEntryUpdateParser } from './felt252DictEntryUpdate';
-import { initSquashDataParser } from './initSquashData';
-import { getCurrentAccessIndexParser } from './getCurrentAccessIndex';
-import { shouldSkipSquashLoopParser } from './shouldSkipSquashLoop';
-import { getCurrentAccessDeltaParser } from './getCurrentAccessDelta';
-import { shouldContinueSquashLoopParser } from './shouldContinueSquashLoop';
-import { getNextDictKeyParser } from './getNextDictKey';
+import { testLessThanParser } from './math/testLessThan';
+import { allocFelt252DictParser } from './dict/allocFelt252Dict';
+import { getSegmentArenaIndexParser } from './dict/getSegmentArenaIndex';
+import { felt252DictEntryInitParser } from './dict/felt252DictEntryInit';
+import { felt252DictEntryUpdateParser } from './dict/felt252DictEntryUpdate';
+import { initSquashDataParser } from './dict/initSquashData';
+import { getCurrentAccessIndexParser } from './dict/getCurrentAccessIndex';
+import { shouldSkipSquashLoopParser } from './dict/shouldSkipSquashLoop';
+import { getCurrentAccessDeltaParser } from './dict/getCurrentAccessDelta';
+import { shouldContinueSquashLoopParser } from './dict/shouldContinueSquashLoop';
+import { getNextDictKeyParser } from './dict/getNextDictKey';
 import { assertLeIsFirstArcExcludedParser } from './assertLeIsFirstArcExcluded';
 import { assertLeFindSmallArcsParser } from './assertLeFindSmallArc';
 import { assertLeIsSecondArcExcludedParser } from './assertLeIsSecondArcExcluded';
